+++ romancal/lib/suffix.py
@@ -82,6 +82,7 @@ _calculated_suffixes = {
     "linearitystep",
     "dark_current",
     "jump",
+    "sourcedetectionstep",
     "tweakregstep",
 }