+++ romancal/source_detection/__init__.py
@@ -1,0 +1,3 @@
+from .source_detection_step import SourceDetectionStep
+
+__all__ = ["SourceDetectionStep"]